chore: move to rigth scope
