+++ packages/homeserver/src/controllers/federation/profiles.controller.ts
@@ -135,41 +135,5 @@ export const profilesPlugin = (app: Elysia) => {
 					description: 'Get event auth for a room',
 				},
 			},
-		)
-		.get(
-			'/_matrix/federation/v1/state_ids/:roomId',
-			({ params, query }) =>
-				profilesService.getStateIds(params.roomId, query.event_id),
-			{
-				params: GetStateIdsParamsDto,
-				query: GetStateIdsQueryDto,
-				response: {
-					200: GetStateIdsResponseDto,
-					400: ErrorResponseDto,
-				},
-				detail: {
-					tags: ['Federation'],
-					summary: 'Get state IDs',
-					description: 'Get state event IDs for a room',
-				},
-			},
-		)
-		.get(
-			'/_matrix/federation/v1/state/:roomId',
-			({ params, query }) =>
-				profilesService.getState(params.roomId, query.event_id!),
-			{
-				params: GetStateParamsDto,
-				query: GetStateQueryDto,
-				response: {
-					200: GetStateResponseDto,
-					400: ErrorResponseDto,
-				},
-				detail: {
-					tags: ['Federation'],
-					summary: 'Get state',
-					description: 'Get state events for a room',
-				},
-			},
 		);
 };
